refactor: combine request_verify and create_account api route
