+++ backend/src/routes/auth/mod.rs
@@ -4,7 +4,6 @@ pub mod login;
 pub mod logout;
 pub mod me;
 pub mod request_reset;
-pub mod request_verify;
 pub mod reset_password;
 pub mod verify;
 pub mod verify_reset;
